refactor: Independently determine if section has data
